chore: restructuring FI profile services
