+++ xroad-catalog-lister/src/main/java/fi/dvv/xroad/catalog/lister/endpoint/services/getorganizations/GetOrganizationsResponseSerializer.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.catalog.lister.endpoint.services.getorganizations;
+package fi.dvv.xroad.catalog.lister.endpoint.services.getorganizations;
 
 import jakarta.xml.soap.SOAPElement;
 import jakarta.xml.soap.SOAPEnvelope;
