+++ xroad-catalog-lister/src/main/java/fi/dvv/xroad/catalog/lister/endpoint/services/hasorganizationchanged/HasOrganizationChangedRequest.java
@@ -22,13 +22,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.catalog.lister.endpoint.services.getcompanies;
+package fi.dvv.xroad.catalog.lister.endpoint.services.hasorganizationchanged;
 
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
+
 @Getter
 @Setter
-public class GetCompaniesRequest {
-    private String businessId;
+public class HasOrganizationChangedRequest {
+    private String guid;
+    private LocalDateTime startDateTime;
+    private LocalDateTime endDateTime;
 }
